reverts cycle duration to an int to fix  incrementing/out of range issues that would cause hard crashes

+++ src/main/java/rs117/hd/HdPluginConfig.java
@@ -662,7 +662,11 @@ public interface HdPluginConfig extends Config
 		position = 2,
 		section = daylightCycleSettings
 	)
-	default double cycleDurationMinutes() {
+	// int, not double: RuneLite's config panel hardcodes the double spinner's step to
+	// 0.1 and ignores @Range for doubles, so a double here makes the arrows increment
+	// by a fraction of a minute and drops the 1-720 clamp. The int spinner steps by 1
+	// and honors @Range.
+	default int cycleDurationMinutes() {
 		return 30;
 	}
